requirements.txt
mysql-connector-python==8.0.33
firebase-admin==5.4.0
google-cloud-firestore==2.10.0
setuptools==67.6.0
